исправил ошибки с хранилищем
