+++ my-app/src/components/LoanCode/LoanCode.module.scss
@@ -23,7 +23,8 @@
     text-align: center;
     width: 3rem;
     height: 3rem;
-    border: 0.0313rem solid #808080;
+    background-color: transparent;
+    border: 0.0313rem solid rgba(128, 128, 128, 1);
     border-radius: 0.5rem;
 
     &:focus {
@@ -44,6 +45,7 @@
     width: 1.25rem;
     height: 1.25rem;
     border: 0.0938rem solid #808080;
+    outline: 0.0938rem solid #808080;
     border-radius: 50%;
   }
 
@@ -53,6 +55,10 @@
 
   &__input:not(:placeholder-shown) + &__inputDecoration {
     display: none;
+  }
+
+  &__input:not(:placeholder-shown) {
+    border: 0.0625rem solid rgba(49, 130, 206, 1);
   }
 
   &__inputError {
