+++ my-app/src/pages/LoanCodePage/LoanCodePage.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import { RootState } from 'src/redux/features/tabs/store';
+import { getStep } from '../../redux/features/store';
 import { LoanCode } from '../../components/LoanCode/LoanCode';
 import { Message } from '../../components/Message/Message';
 
@@ -8,7 +8,7 @@ import classes from './LoanCodePage.module.scss';
 import congratulations from '../../assets/images/offerBox.webp';
 
 export const LoanCodePage = () => {
-  const { activeStep } = useSelector((state: RootState) => state.steps);
+  const { activeStep } = useSelector(getStep);
 
   return (
     <main className={classes.main} aria-labelledby="loan-code-title">
